Remove auto clone of cassandra git repo
Originally the project required building Cassandra to use it.  This
proved to be quite cumbesome.  It's still possible (and will always be
possible) to do custom builds, but I think it's unnecessary for the most
part.  We definitely don't need to clone a repo in the background to get
started using this tool.

Closes #43