Fixes broken unit tests that were using the interactive process
When I had moved the public / private keys from a manual process to an automatic one,
I accidentally broke some tests that now have a dependency on a user profile existing.

The solution here is to create the user profile as part of the test context. we don't
need real data here, we just need it to not try to do the interactive setup.

Fixes #66